% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fract_curve_clust.R
\name{fract_curve_clust}
\alias{fract_curve_clust}
\title{Decide the number of clusters by the fracture point of the height.}
\usage{
fract_curve_clust(df.features = OTUdata_all,
  df.phenotype = pData(obj.ADS), var.phenoGroup = "Disease",
  method.dist.row = "manhattan", method.dist.col = "manhattan",
  method.hclust.row = "ward", method.hclust.col = "ward",
  fisher_test = TRUE, dir.output = NULL, get.df_of_IYs,
  fn.plot_pdf = NULL, fn.df_of_IYs = NULL)
}
\arguments{
\item{df.features}{A data.frame-class object.}

\item{df.phenotype}{A data.frame-class object.}

\item{method.dist.row}{A distance metric for clustering of features.}

\item{method.dist.col}{A distance metric for clustering of samples.}

\item{method.hclust.row}{An aggromerizing algorithm for clustering of features.}

\item{method.hclust.col}{An aggromerizing algorithm for clustering of samples.}

\item{dir.output}{The directory to output results.}

\item{get.df_of_IYs}{A logical if IYs will be obtained as data.frame in result.}

\item{fn.plot_pdf}{The pdf file name for outputs.}

\item{fn.df_of_IYs}{The csv file name for IY output.}
}
\description{
Decide the number of clusters by the fracture point of the height.
}
